refactor : added error statements and refactored the controller to take username in request

+++ controller/managerController.js
@@ -8,22 +8,24 @@ const managerController = {
      * @route POST/manager/addStatus
      */
     addStatus: async (req, res) => {
-        const { userId } = req.body;
-        if (!userId) {
-            return res.status(400).json({ message: "User ID is required" });
+        const { username} = req.body;
+        if (!username) {
+            return res.status(400).json({ message: "Username is required" });
         }
 
-        const isUserExist = await userProfile.findOne({ userId });
+        const isUserExist = await userProfile.findOne({ username });
         if (!isUserExist) {
             return res.status(404).json({ message: "User not found" });
         }
-
+        if (isUserExist.isManager) {
+            return res.status(400).json({ message: "User already has manager status" });
+        }
         try {
-            await userProfile.findOneAndUpdate({ userId: userId }, { $set: { isManager: true } }, { upsert: true });
+            await userProfile.findOneAndUpdate({ username }, { $set: { isManager: true } }, { upsert: true });
 
             // create manager record
             const managerRecord = new manager({
-                managerId: userId,
+                managerId: isUserExist._id,
 
             });
             await managerRecord.save();
@@ -43,13 +45,22 @@ const managerController = {
     **/
 
     removeStatus: async (req, res) => {
-        const { managerId } = req.body;
+        const { username } = req.body;
+        const user = await userProfile.findOne({ username });
+        if (!user) {
+            return res.status(404).json({ message: "User not found" });
+        }
+
+        if (!user.isManager) {
+            return res.status(400).json({ message: "User does not have manager status" });
+        }
+        const managerId = user._id;
         if (!managerId) {
             return res.status(400).json({ message: "Manager ID is required" });
         }
         try {
             await manager.findOneAndDelete({ managerId: managerId });
-            await userProfile.findOneAndUpdate({ userId: managerId }, { $set: { isManager: false } });
+            await userProfile.findOneAndUpdate({ _id: managerId }, { $set: { isManager: false } });
             res.status(200).json({ message: "Manager status removed successfully" });
         } catch (error) {
             console.error("Error removing manager status from user:", error);
@@ -63,15 +74,24 @@ const managerController = {
      * @route POST/manager/addClient
      */
     addClient: async (req, res) => {
-        const { managerId, clientUsername } = req.body;
+        const { username, clientUsername } = req.body;
+        if (!username ||!clientUsername) {
+            return res.status(400).json({ message: "Username and clientUsername are required" });
+        }
+        const user = await userProfile.findOne({ username });
+        if (!user.isManager) {
+            return res.status(400).json({ message: "User does not have manager status" });
+        }
 
-        //find client by username 
+        const managerId = user._id;
+        
+     
         const client = await userProfile.findOne({ username: clientUsername });
         if (!client) {
             return res.status(404).json({ message: "Client not found" });
         }
         try {
-            await manager.findOneAndUpdate({ managerId: managerId }, { $push: { clientId: client.userId } });
+            await manager.findOneAndUpdate({ managerId: managerId }, { $push: { clientId: client._id} });
             res.status(200).json({ message: "Client added successfully" });
         } catch (error) {
             console.error("Error adding client to manager:", error);
@@ -84,13 +104,21 @@ const managerController = {
      * @route POST /manager/removeClient
      */
     removeClient: async (req, res) => {
-        const { managerId, clientUsername } = req.body;
+        const { username, clientUsername } = req.body;
+        if (!username ||!clientUsername) {
+            return res.status(400).json({ message: "Username and clientUsername are required" });
+        }
+        const user = await userProfile.findOne({ username });
+        if (!user.isManager) {
+            return res.status(400).json({ message: "User does not have manager status" });
+        }
+        const managerId = user._id;
         const client = await userProfile.findOne({ username: clientUsername }); 
         if (!client) {
             return res.status(404).json({ message: "Client not found" });
         }
         try {
-            await manager.findOneAndUpdate({ managerId: managerId }, { $pull: { clientId: client.userId } });
+            await manager.findOneAndUpdate({ managerId: managerId }, { $pull: { clientId: client._id} });
             res.status(200).json({ message: "Client removed successfully" });
         } catch (error) {
             console.error("Error removing client from manager:", error);
@@ -105,11 +133,12 @@ const managerController = {
 
     updateManager: async (req, res) => {
         const updateFields = req.body;
-        if (!updateFields.managerId) {
-            return res.status(400).json({ message: "Is not a manager" });
+        const user = await userProfile.findOne({ username: updateFields.username });
+        if (!user.isManager) {
+            return res.status(400).json({ message: " User is not a manager" });
         }
         try {
-            await manager.findOneAndUpdate({ managerId: updateFields.managerId }, updateFields);
+            await manager.findOneAndUpdate({ managerId:user._id }, updateFields);
             res.status(200).json({ message: "Manager details updated successfully" });
         } catch (error) {
             console.error("Error updating manager details:", error);
@@ -146,7 +175,7 @@ const managerController = {
                 return res.status(404).json({ message: "User not found" });
             }
 
-            const managerId = user.userId;
+            const managerId = user._id;
             if (!user) {
                 return res.status(404).json({ message: "Manager not found" });
             }
